refactor: RestaurantCards의 하나의 Restaurant을 RestaurantCard 컴포넌트로 분리
